requirements.txt
torch
pandas
scikit-learn
nltk
python-docx
matplotlib
